♻️ [Refactoring]: #041 objects/ 配下を 1機能1フォルダ化 + helper パス更新
- objects/object-store: top-level tests と test.helpers を object-store/__tests__/ へ集約
- objects/object-store/entry-readers/{inline,object-stream}: 個別フォルダ化
- objects/object-stream-extractor/body: top-level tests と helpers を body/__tests__/ へ集約
- objects/object-parser: top-level tests を object-parser/__tests__/ へ集約（Task A から漏れていた分）
- 主実装内の相対 import を +1、test 内を +2 階層調整
- cross-feature helper 参照 (object-store / object-stream-body) のパスを __tests__/ 経由に更新
- types.ts は据え置き (object-parser, object-store, object-stream-extractor)
- Test Files=124 / Tests=1688 維持
