+++ packages/core/src/objects/object-parser/__tests__/object-parser.facade-helpers.test.ts
@@ -1,6 +1,6 @@
 import { assert, expect, test } from "vitest";
-import { ByteOffset } from "../../pdf/types/byte-offset/index";
-import { ObjectParser } from "./index";
+import { ByteOffset } from "../../../pdf/types/byte-offset/index";
+import { ObjectParser } from "../index";
 
 const enc = (s: string): Uint8Array => new TextEncoder().encode(s);
