+++ packages/core/src/objects/object-stream-extractor/dict/__tests__/object-stream-dict.parse.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test } from "vitest";
-import { makeDict } from "../../body/object-stream-body.test.helpers";
+import { makeDict } from "../../body/__tests__/object-stream-body.test.helpers";
 import { ObjectStreamDict } from "../index";
 
 test("parseは正しい辞書から/First,/N,needsDecompressを取得できる", () => {
