+++ packages/core/src/objects/object-stream-extractor/header/__tests__/object-stream-header.validation.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test } from "vitest";
-import { enc } from "../../body/object-stream-body.test.helpers";
+import { enc } from "../../body/__tests__/object-stream-body.test.helpers";
 import { ObjectStreamHeader } from "../index";
 
 test("parseは1つのobjNum/offsetペアを持つヘッダをパースできる", () => {
